edit home.html, idFind.html, logins.html, pwdFind.html, register.html, register.html, url.py, views.py, views_identify.py, views_market.py
