+++ templates/home.html
@@ -56,12 +56,12 @@
 
       <nav id="navbar" class="navbar">
         <ul>
-          <li><a class="nav-link scrollto active" href="#hero">Home</a></li>
+          <li><a class="nav-link scrollto active" href="/">Home</a></li>
           {% if request.session.sessionid == None %}
-          <li><a class="nav-link scrollto" href="/login">login</a></li>
-          <li><a class="nav-link scrollto" href="/register">register</a></li>
+          <li><a class="nav-link scrollto" href="/identify/login">login</a></li>
+          <li><a class="nav-link scrollto" href="/identify/register">register</a></li>
           {% else %}
-          <li><a class="nav-link scrollto" href="/logout">logout</a></li>
+          <li><a class="nav-link scrollto" href="/identify/logout">logout</a></li>
           {% endif %}
           <li><a class="nav-link scrollto" href="#services">categori</a></li>
           <li><a class="nav-link scrollto " href="#portfolio">popular market</a></li>
@@ -192,37 +192,37 @@
           <div class="col-lg-3 col-md-3 col-6 col-6">
             <div class="icon-box">
               <i class="ri-store-line" style="color: #ffbb2c;"></i>
-              <h3><a href="/chicken">치킨</a></h3>
+              <h3><a href="/market/chicken">치킨</a></h3>
             </div>
           </div>
           <div class="col-lg-3 col-md-3 col-6">
             <div class="icon-box">
               <i class="ri-bar-chart-box-line" style="color: #5578ff;"></i>
-              <h3><a href="/chinese">중식</a></h3>
+              <h3><a href="/market/chinese">중식</a></h3>
             </div>
           </div>
           <div class="col-lg-3 col-md-3 col-6 mt-4 mt-md-0">
             <div class="icon-box">
               <i class="ri-calendar-todo-line" style="color: #e80368;"></i>
-              <h3><a href="/dessert">디저트</a></h3>
+              <h3><a href="/market/dessert">디저트</a></h3>
             </div>
           </div>
           <div class="col-lg-3 col-md-3 col-6 mt-4 mt-lg-0">
             <div class="icon-box">
               <i class="ri-paint-brush-line" style="color: #e361ff;"></i>
-              <h3><a href="/fastfood">패스트푸드</a></h3>
+              <h3><a href="/market/fastfood">패스트푸드</a></h3>
             </div>
           </div>
           <div class="col-lg-3 col-md-3 col-6 mt-4">
             <div class="icon-box">
               <i class="ri-database-2-line" style="color: #47aeff;"></i>
-              <h3><a href="/pizza">피자</a></h3>
+              <h3><a href="/market/pizza">피자</a></h3>
             </div>
           </div>
           <div class="col-lg-3 col-md-3 col-6 mt-4">
             <div class="icon-box">
               <i class="ri-gradienter-line" style="color: #ffa76e;"></i>
-              <h3><a href="/western">미국 음식</a></h3>
+              <h3><a href="/market/western">미국 음식</a></h3>
             </div>
           </div>
         </div>
