+++ templates/identify/idFind.html
@@ -64,9 +64,9 @@
       <div class="container">
 
         <ol>
-          <li><a href="/">Home</a></li>
-          <li><a href="/login">Login</a></li>
-          <li><a href="/register">register</a></li>
+          <li><a href="/identify/">Home</a></li>
+          <li><a href="/identify/login">Login</a></li>
+          <li><a href="/identify/register">register</a></li>
         </ol>
 
       </div>
