+++ templates/identify/logins.html
@@ -49,7 +49,8 @@
       <nav id="navbar" class="navbar">
         <ul>
           <li><a class="nav-link scrollto " href="/">Home</a></li>
-          <li><a class="active" href="/register">Register</a></li>
+          <li><a class="active" href="/login">Login</a></li>
+          <li><a class="nav-link scrollto" href="/identify/register">Register</a> </li>
         </ul>
         <i class="bi bi-list mobile-nav-toggle"></i>
       </nav><!-- .navbar -->
@@ -64,10 +65,11 @@
       <div class="container">
 
         <ol>
-          <li><a href="index.html">Home</a></li>
-          <li>Register</li>
+          <li><a href="/">Home</a></li>
+          <li><a href="/identify/login">Login</a></li>
+          <li><a href="/identify/register">register</a></li>
         </ol>
-        <h2>register</h2>
+        <h2>login</h2>
 
       </div>
     </section><!-- End Breadcrumbs -->
@@ -83,31 +85,23 @@
             <article class="entry">
 
               <h2 class="entry-title">
-                <a href="blog-single.html">register</a>
+                <a href="blog-single.html">login</a>
               </h2>
 
               <div class="entry-content">
-            <form action="/registerimpl" method="post" role="form" class="php-email-form">
-              {% csrf_token %}
-              <div class="form-group mt-3">
-                <input type="text" class="form-control" name="id" id="id" placeholder="Id" required>
-              </div>
-              <div class="form-group mt-3">
-                <input type="text" class="form-control" name="password" id="password" placeholder="Password" required>
-              </div>
-              <div class="form-group mt-3">
-                <input type="text" class="form-control" name="name" id="name" placeholder="Name" required>
-              </div>
-              <div class="form-group mt-3">
-                <input type="text" class="form-control" name="address" id="address" placeholder="Address" required>
-              </div>
-              <div  class="form-group mt-3">
-                <input type="text" class="form-control" name="email" id="email" placeholder="Email" required>
-              </div>
-              <br>
-              <div style="display: inline-block" class="text-center"><button type="submit">register</button></div>
-              <div style="display: inline-block" class="text-center"><button type="reset">reset</button></div>
-            </form>
+                <form action="/loginimpl" method="post" role="form" class="php-email-form">
+                  {% csrf_token %}
+                  <div class="form-group mt-3">
+                    <input type="text" class="form-control" name="id" id="id" placeholder="Id" required>
+                  </div>
+                  <div class="form-group mt-3">
+                    <input type="text" class="form-control" name="password" id="password" placeholder="Password" required>
+                  </div>
+                  <br>
+                  <div style="display: inline-block" class="text-center"><button type="submit">login</button></div><br>
+                </form>
+                <div style="display: inline-block" class="text-center"><button><a href="/idfind">idfind</a></button></div>
+                <div style="display: inline-block" class="text-center"><button><a href="/pwdfind">pwdfind</a></button></div>
               </div>
             </article>
           </div>
@@ -199,7 +193,6 @@
   <script src="{% static 'assets/vendor/glightbox/js/glightbox.min.js' %}"></script>
   <script src="{% static 'assets/vendor/isotope-layout/isotope.pkgd.min.js' %}"></script>
   <script src="{% static 'assets/vendor/swiper/swiper-bundle.min.js' %}"></script>
-
 
   <!-- Template Main JS File -->
   <script src="{% static 'assets/js/main.js' %}"></script>
